+++ src/apiConfig.ts
@@ -43,6 +43,48 @@ export const getAPIVersionForPath = () => {
   return 'v' + getAPIVersion().replace('.', '_');
 };
 
+/**
+ * Gets the configured clientId for the OCAPI calls from the VSCode setting, or
+ * uses the default if none is set.
+ * @return {string} - Returns the 30 character client Id configured for
+ *    making calls to OCAPI.
+ */
+export const getClientId = () => {
+  let clientId = apiConfig.clientId;
+  const workspaceConfig: WorkspaceConfiguration = workspace.getConfiguration(
+    'extension.sfccmetadata'
+  );
+  const configId = String(workspaceConfig.get('ocapi.clientid'));
+  const tester = /^[\S]{30}$/;
+  if (configId && !tester.test(configId)) {
+    window.showErrorMessage('Value configured for OCAPI clientId is not valid: ' +
+      configId + '. Using default version API client Id.');
+  }
+
+  return configId && tester.test(configId) ? configId : clientId;
+};
+
+/**
+ * Gets the configured client password for the OCAPI calls from VSCode setting,
+ * or uses the default if none is set.
+ * @return {string} - Returns the 30 character client Id configured for
+ *    making calls to OCAPI.
+ */
+export const getClientPass = () => {
+  let defaultPass = apiConfig.clientId;
+  const workspaceConfig: WorkspaceConfiguration = workspace.getConfiguration(
+    'extension.sfccmetadata'
+  );
+  const configPass = String(workspaceConfig.get('ocapi.clientpassword'));
+  const tester = /^[\S]{30}$/;
+  if (configPass && !tester.test(configPass)) {
+    window.showErrorMessage('Value configured for OCAPI client password is not valid: ' +
+      configPass + '. Using default version API client Id.');
+  }
+
+  return configPass && tester.test(configPass) ? configPass : defaultPass;
+};
+
 /* ========================================================================
  * Exported API Configuration Object
  * ======================================================================== */
@@ -99,7 +141,7 @@ export const apiConfig = {
               use: 'PATH_PARAMETER'
             }
           ],
-          path: '/custom_object_definitions/{objectType}/attribute_definitions'
+          path: 'custom_object_definitions/{objectType}/attribute_definitions'
         },
 
         /* ==================================================================
@@ -122,6 +164,141 @@ export const apiConfig = {
             }
           ],
           path: 'custom_object_definitions/{objectType}/attribute_definitions'
+        }
+      }
+    },
+
+    /***************************************************************************
+     * OCAPI : Data API
+     * Resource : Jobs
+     **************************************************************************/
+    jobs: {
+      api: 'data',
+      availableCalls: {
+        /* ==================================================================
+         * POST Execute Job
+         * ================================================================== */
+        executeJob: {
+          authorization: 'BM_USER',
+          headers: { 'Content-Type': 'application/json' },
+          method: 'POST',
+          params: [
+            {
+              id: 'job_id',
+              type: 'string',
+              use: 'PATH_PARAMETER'
+            }
+          ],
+          path:
+            'jobs/{job_id}/executions'
+        },
+
+        /* ==================================================================
+         * GET Job Execution
+         * ================================================================== */
+        getExecution: {
+          authorization: 'BM_USER',
+          headers: { 'Content-Type': 'application/json' },
+          method: 'GET',
+          params: [
+            {
+              id: 'job_id',
+              type: 'string',
+              use: 'PATH_PARAMETER'
+            },
+            {
+              id: 'execution_id',
+              type: 'string',
+              use: 'PATH_PARAMETER'
+            }
+          ],
+          path:
+            'jobs/{job_id}/executions/{execution_id}'
+        }
+      }
+    },
+
+    /***************************************************************************
+     * OCAPI : Data API
+     * Resource : SitePreferences
+     **************************************************************************/
+    sitePreferences: {
+      api: 'data',
+      availableCalls: {
+        /* ==================================================================
+         * GET SitePreference by Id
+         * ================================================================== */
+        getPreference: {
+          authorization: 'BM_USER',
+          headers: { 'Content-Type': 'application/json' },
+          method: 'GET',
+          params: [
+            {
+              id: 'groupId',
+              type: 'string',
+              use: 'PATH_PARAMETER'
+            },
+            {
+              id: 'instanceType',
+              type: 'string',
+              use: 'PATH_PARAMETER'
+            },
+            {
+              id: 'preferenceId',
+              type: 'string',
+              use: 'PATH_PARAMETER'
+            }
+          ],
+          path:
+            'site_preferences/preference_groups/{groupId}/{instanceType}/preferences/{preferenceId}'
+        }
+      }
+    },
+
+    /***************************************************************************
+     * OCAPI : Data API
+     * Resource : Sites
+     **************************************************************************/
+    sites: {
+      api: 'data',
+      availableCalls: {
+        /* ==================================================================
+         * GET ALL SITES
+         * ================================================================== */
+        getAll: {
+          authorization: 'BM_USER',
+          headers: { 'Content-Type': 'application/json' },
+          method: 'GET',
+          params: [],
+          path: '/sites'
+        },
+
+        /* ==================================================================
+         * PATCH Site Preferece value
+         * ================================================================== */
+        setPrefValue: {
+          authorization: 'BM_USER',
+          headers: { 'Content-Type': 'application/json' },
+          method: 'PATCH',
+          params: [
+            {
+              id: 'site_id',
+              type: 'string',
+              use: 'PATH_PARAMETER'
+            },
+            {
+              id: 'instance_type',
+              type: 'string',
+              use: 'PATH_PARAMETER'
+            },
+            {
+              id: 'group_id',
+              type: 'string',
+              use: 'PATH_PARAMETER'
+            }
+          ],
+          path:
+            'sites/{site_id}/site_preferences/preference_groups/{group_id}/{instance_type}'
         }
       }
     },
@@ -408,87 +585,6 @@ export const apiConfig = {
           ],
           path:
             'system_object_definitions/{objectType}/attribute_groups/{groupId}/attribute_definitions/{attributeId}'
-        }
-      }
-    },
-
-    /***************************************************************************
-     * OCAPI : Data API
-     * Resource : SitePreferences
-     **************************************************************************/
-    sitePreferences: {
-      api: 'data',
-      availableCalls: {
-        /* ==================================================================
-         * GET SitePreference by Id
-         * ================================================================== */
-        getPreference: {
-          authorization: 'BM_USER',
-          headers: { 'Content-Type': 'application/json' },
-          method: 'GET',
-          params: [
-            {
-              id: 'groupId',
-              type: 'string',
-              use: 'PATH_PARAMETER'
-            },
-            {
-              id: 'instanceType',
-              type: 'string',
-              use: 'PATH_PARAMETER'
-            },
-            {
-              id: 'preferenceId',
-              type: 'string',
-              use: 'PATH_PARAMETER'
-            }
-          ],
-          path:
-            '/site_preferences/preference_groups/{groupId}/{instanceType}/preferences/{preferenceId}'
-        }
-      }
-    },
-
-    /***************************************************************************
-     * OCAPI : Data API
-     * Resource : Sites
-     **************************************************************************/
-    sites: {
-      api: 'data',
-      availableCalls: {
-        /* ==================================================================
-         * GET all sites.
-         * ================================================================== */
-        getAll: {
-          authorization: 'BM_USER',
-          headers: { 'Content-Type': 'application/json' },
-          method: 'GET',
-          params: [],
-          path: '/sites'
-        },
-        setPrefValue: {
-          authorization: 'BM_USER',
-          headers: { 'Content-Type': 'application/json' },
-          method: 'PATCH',
-          params: [
-            {
-              id: 'site_id',
-              type: 'string',
-              use: 'PATH_PARAMETER'
-            },
-            {
-              id: 'instance_type',
-              type: 'string',
-              use: 'PATH_PARAMETER'
-            },
-            {
-              id: 'group_id',
-              type: 'string',
-              use: 'PATH_PARAMETER'
-            }
-          ],
-          path:
-            'sites/{site_id}/site_preferences/preference_groups/{group_id}/{instance_type}'
         }
       }
     }