+++ src/documents/PreferenceValue.ts
@@ -7,7 +7,6 @@
 import { IOCAPITypes } from '../interfaces/IOCAPITypes';
 import ObjectAttributeDefinition from './ObjectAttributeDefinition';
 
-
 /**
  * @class
  * @classdesc - A class for handling the Sites API document for recieving &
@@ -18,7 +17,7 @@ export default class PreferenceValue {
   public description: IOCAPITypes.ILocalizedString = { default: '' };
   public displayName: IOCAPITypes.ILocalizedString = { default: '' };
   public attributeDefinition: ObjectAttributeDefinition;
-  public siteValues: Map<string, object> = new Map();
+  public siteValues: object = {};
   public valueType: string = '';
 
   /**